NPC Updates

+++ Assets/_Project/_Scripts/NPC/StateMachine/NPC_InvisibleState.cs
@@ -18,7 +18,7 @@ public class NPC_InvisibleState : NPC_BaseState
     public override void FixedUpdateState(){}
     public override void ExitState(){}
     public override void CheckSwitchStates(){
-        if(Ctx.NPC.currentScene == SceneManager.GetActiveScene().buildIndex){
+        if((int)Ctx.NPC.currentScene == SceneManager.GetActiveScene().buildIndex){
             SwitchStates(Factory.GetBaseState(NPC_StateNames.Visible));
         }
     }
