+++ Assets/_Project/_Scripts/NPC/StateMachine/NPC_VisibleState.cs
@@ -21,7 +21,7 @@ namespace MyTownProject.NPC
         public override void ExitState() { }
         public override void CheckSwitchStates()
         {
-            if (Ctx.NPC.currentScene != SceneManager.GetActiveScene().buildIndex)
+            if ((int)Ctx.NPC.currentScene != SceneManager.GetActiveScene().buildIndex)
             {
                 SwitchStates(Factory.GetBaseState(NPC_StateNames.Invisible));
             }
